Interpol: Move helper methods to main interpolation file
